refactor(api): trim style discovery helpers — inline link builders, drop hand-typed doc shapes
